feat(TASK-741): Extract interfaces from services
