+++ spring-data-jpa/src/main/java/ru/sabitov/example/service/AuthorService.java
@@ -1,18 +1,8 @@
 package ru.sabitov.example.service;
 
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
 import ru.sabitov.example.dto.AuthorDto;
 import ru.sabitov.example.dto.CreateAuthorDto;
-import ru.sabitov.example.mapper.AuthorMapper;
-import ru.sabitov.example.repository.AuthorRepository;
 
-@Service
-@RequiredArgsConstructor
-public class AuthorService {
-    private final AuthorRepository authorRepository;
-
-    public AuthorDto create(CreateAuthorDto dto) {
-        return AuthorMapper.toDto(authorRepository.save(AuthorMapper.toEntity(dto)));
-    }
+public interface AuthorService {
+    AuthorDto create(CreateAuthorDto dto);
 }
