+++ spring-data-jpa/src/main/java/ru/sabitov/example/service/AuthorServiceImpl.java
@@ -1,0 +1,19 @@
+package ru.sabitov.example.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import ru.sabitov.example.dto.AuthorDto;
+import ru.sabitov.example.dto.CreateAuthorDto;
+import ru.sabitov.example.mapper.AuthorMapper;
+import ru.sabitov.example.repository.AuthorRepository;
+
+@Service
+@RequiredArgsConstructor
+public class AuthorServiceImpl implements AuthorService {
+    private final AuthorRepository authorRepository;
+
+    @Override
+    public AuthorDto create(CreateAuthorDto dto) {
+        return AuthorMapper.toDto(authorRepository.save(AuthorMapper.toEntity(dto)));
+    }
+}
